Remove old Error type
By now we have adjusted all users of crate-level errors to use the new
Error type instead of the old one. With this change we conclude the
switch by removing the old type altogether.
